requirements.txt
# Delta-Neutral Trading Bot Dependencies
# Python 3.10+ required

# HTTP/API clients
httpx>=0.25.0
aiohttp>=3.9.0
websockets>=12.0

# Cryptography (Starknet signatures)
starknet-py>=0.21.0
eth-account>=0.10.0
eth-utils>=2.3.0
pycryptodome>=3.19.0

# Data processing
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Async utilities
asyncio-throttle>=1.0.2
tenacity>=8.2.0

# Logging & monitoring
structlog>=24.1.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.8.0

# Code quality
black>=24.1.0
ruff>=0.1.0

# Hyperliquid SDK (TradeXYZ)
hyperliquid-python-sdk @ git+https://github.com/hyperliquid-dex/hyperliquid-python-sdk.git

# Utilities
python-dateutil>=2.8.0
